split original header.html file
Removed every code in header.html that was not the navbar and fixed the rest of the code in a home.html layout and page. This file will contain all parts of the homepage as opposed to having header.html for the top of the site and another page for the rest of the homepage.

This way, if another page is ever created, it will be easy to reuse the nav.